Factour out spectral scaler in its own file
